+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvDataSource.java
@@ -66,7 +66,7 @@ public abstract class CsvDataSource {
    */
   @Deprecated private boolean notYetLoggedWarning = true;
 
-  public CsvDataSource(String csvSep, String folderPath, FileNamingStrategy fileNamingStrategy) {
+  protected CsvDataSource(String csvSep, String folderPath, FileNamingStrategy fileNamingStrategy) {
     this.csvSep = csvSep;
     this.connector = new CsvFileConnector(folderPath, fileNamingStrategy);
   }